% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deposit}
\alias{deposit}
\title{deposit}
\format{A data frame with three variables: \code{date}, \code{dp.dd},
  \code{dp.td}. \code{date} starts from Fiscal year 2009, that is July 1, 2009 and ends on June 01, 2020. \code{dp.dd} represents demand deposits and \code{dp.td} represents time deposits.}
\usage{
deposit
}
\description{
deposit includes data on demand deposits
}
\keyword{datasets}
