% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{money}
\alias{money}
\title{money}
\format{A data frame with three variables: \code{date}, \code{M1},
  \code{M1}. \code{date} starts from Fiscal year 2009, that is July 1, 2009 and ends on April 01, 2019. \code{M1} \code{M1} variables are self explanatory}
\usage{
money
}
\description{
money includes data on M1 and M2
}
\keyword{datasets}
