% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deposit}
\alias{deposit}
\title{Macro BD}
\format{There are two data frames
deposit 

deposit includes data on demand deposits}
\usage{
deposit
}
\description{
macrobd package includes data from economic trends of Bangladesh
}
\keyword{datasets}
